+++ packages/sst/src/cli/commands/build.ts
@@ -1,3 +1,4 @@
+import { useProject } from "../../project.js";
 import { Program } from "../program.js";
 import { createSpinner } from "../spinner.js";
 
@@ -5,14 +6,13 @@ export const build = (program: Program) =>
   program.command(
     "build",
     "Build stacks code",
-    yargs => yargs.option("from", { type: "string" }),
+    (yargs) => yargs.option("from", { type: "string" }),
     async () => {
       const { Stacks } = await import("../../stacks/index.js");
       const spinner = createSpinner("Building stacks").start();
-      const fn = await Stacks.build();
       await Stacks.synth({
-        fn,
-        mode: "deploy"
+        fn: useProject().stacks,
+        mode: "deploy",
       });
       spinner.succeed();
       process.exit(0);